+++ NEWS.md
@@ -1,6 +1,10 @@
 
 ### acefile 0.5.2-dev
 
+-   Renamed AceFile to AceArchive, but AceFile is still available as an alias.
+-   Hidden AceInfo class from the API, it is still there but not in __all__.
+-   Added all exceptions to the API.
+
 
 ### acefile 0.5.1 2017-07-02